requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
discord==1.7.3
discord.py==1.7.3
idna==3.3
multidict==6.0.2
python-dotenv==0.20.0
requests==2.28.1
soupsieve==2.3.2.post1
typing-extensions==4.3.0
urllib3==1.26.11
yarl==1.8.1